pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-runtime-lab"
version = "0.1.0"
description = "Governance-first agent runtime lab with Planner/Executor/Critic and dual execution loops"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "laoyu17" }
]
dependencies = [
  "pydantic>=2.7,<3",
  "PyYAML>=6.0,<7"
]

[project.optional-dependencies]
dev = [
  "black>=24.10,<25",
  "mypy>=1.11,<2",
  "pytest>=8.3,<9",
  "pytest-cov>=5.0,<6",
  "ruff>=0.6,<1"
]

[project.scripts]
agent-runtime-lab = "agent_runtime_lab.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py311"]
extend-exclude = "codex_middleware/"

[tool.ruff]
line-length = 88
target-version = "py311"
src = ["src", "tests"]
extend-exclude = ["codex_middleware"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]

[tool.mypy]
python_version = "3.11"
files = ["src"]
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true

[tool.pytest.ini_options]
addopts = "-q --cov=src --cov-report=term-missing --cov-fail-under=80"
testpaths = ["tests"]

[tool.coverage.run]
omit = [
  "src/agent_runtime_lab/cli.py"
]
